misc: change to be func of struct
