+++ src/cli/exec.rs
@@ -19,7 +19,6 @@ use rattler_virtual_packages::VirtualPackage;
 use reqwest_middleware::ClientWithMiddleware;
 
 use crate::prefix::Prefix;
-use pixi_config::gateway::from_pixi_config;
 use pixi_config::{self, Config, ConfigCli};
 use pixi_progress::{await_in_progress, global_multi_progress, wrap_in_progress};
 use pixi_utils::{reqwest::build_reqwest_clients, PrefixGuard};
@@ -157,7 +156,7 @@ pub async fn create_exec_prefix(
     let gateway = Gateway::builder()
         .with_cache_dir(cache_dir.join("repodata"))
         .with_client(client.clone())
-        .with_channel_config(from_pixi_config(config))
+        .with_channel_config(config.into())
         .finish();
 
     // Determine the specs to use for the environment
