+++ src/cli/global/install.rs
@@ -27,7 +27,7 @@ use reqwest_middleware::ClientWithMiddleware;
 
 use super::common::{channel_name_from_prefix, find_designated_package, BinDir, BinEnvDir};
 use crate::{cli::has_specs::HasSpecs, prefix::Prefix};
-use pixi_config::{self, gateway::new_gateway, Config, ConfigCli};
+use pixi_config::{self, Config, ConfigCli};
 use pixi_progress::{await_in_progress, global_multi_progress, wrap_in_progress};
 
 /// Installs the defined package in a global accessible location.
@@ -315,7 +315,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
     // Fetch the repodata
     let (_, auth_client) = build_reqwest_clients(Some(&config));
 
-    let gateway = new_gateway(auth_client.clone(), config.clone());
+    let gateway = config.gateway(auth_client.clone());
 
     let repodata = gateway
         .query(
