+++ src/project/repodata.rs
@@ -1,6 +1,5 @@
 use crate::project::Project;
 
-use pixi_config::gateway::from_pixi_config;
 use rattler_repodata_gateway::Gateway;
 use std::path::PathBuf;
 
@@ -18,7 +17,7 @@ impl Project {
             Gateway::builder()
                 .with_client(self.authenticated_client().clone())
                 .with_cache_dir(cache_dir.join("repodata"))
-                .with_channel_config(from_pixi_config(&self.config))
+                .with_channel_config(self.config().into())
                 .finish()
         })
     }
